1_sportprog/40/requirements.txt
discord.py==1.7.3
python-decouple==3.6
